feat: add sqlite memory
