+++ thinagents/core/__init__.py
@@ -8,6 +8,6 @@ Key components include tool definition and schema generation capabilities.
 from thinagents.core.tool import tool
 from thinagents.core.agent import Agent
 from thinagents.core.response_models import ThinagentResponse
-from thinagents.core.memory import BaseMemory, InMemoryStore, FileMemory, ConversationInfo
+from thinagents.memory import BaseMemory, InMemoryStore, FileMemory, ConversationInfo
 
 __all__ = ["tool", "Agent", "ThinagentResponse", "BaseMemory", "InMemoryStore", "FileMemory", "ConversationInfo"]
